[EA3-183] refactor : 시간 투표 검증 로직을 TimeVoteValidator 클래스로 분리
